Add several test runs, add an invalid typst file for test use.  This will cover main loop refactor.

+++ tests.py
@@ -16,10 +16,9 @@ def remove_glob(pattern: str):
     '''
     base = Path() # set root path for the glob
     for path in base.glob(pattern):
-        if path.is_file():
-            path.unlink(missing_ok=True)
+        if path.is_file(): path.unlink(missing_ok=True)
 
-class TestAddFunction(unittest.TestCase):
+class TestTypstAction(unittest.TestCase):
 
     # store path to python interpreter.
     python_path = 'python3'
@@ -39,23 +38,52 @@ class TestAddFunction(unittest.TestCase):
         self.assertTrue(Path.exists(Path('tests/math.pdf')))
         self.assertFalse(Path.exists(Path('tests/valid.pdf')))
 
-    def test_run_with_directory_as_arg(self):
-        # Compile a file using source file argument, passing in a directory name. 
+    def test_run_with_wildcard_arg(self):
+        # Compile a file using source file argument, passing in a file name.
         remove_glob('tests/*.pdf')          # clean up any existing output artifacts first
-        command = [self.python_path, 'entrypoint.py', 'tests', '', '']
+        command = [self.python_path, 'entrypoint.py', 'tests/*.typ', '', '']
         result = subprocess.run(command, capture_output=True, text=True)
         self.assertEqual(result.returncode,0)
         self.assertTrue(Path.exists(Path('tests/math.pdf')))
         self.assertTrue(Path.exists(Path('tests/valid.pdf')))
 
+    def test_run_with_multiline_source_file_arg(self):
+        # Compile a file using source file argument, passing in a file name.
+        remove_glob('tests/*.pdf')          # clean up any existing output artifacts first
+        command = [self.python_path, 'entrypoint.py', 'tests/math.typ\ntests/valid.typ', '', '']
+        result = subprocess.run(command, capture_output=True, text=True)
+        self.assertEqual(result.returncode,0)
+        self.assertTrue(Path.exists(Path('tests/math.pdf')))
+        self.assertTrue(Path.exists(Path('tests/valid.pdf')))
+
+    def test_run_with_directory_as_arg(self):
+        # Compile a file using source file argument, passing in a directory name. 
+        # Note that the directory contains an invalid test file, so returncode will be 1.
+        remove_glob('tests/*.pdf')          # clean up any existing output artifacts first
+        command = [self.python_path, 'entrypoint.py', 'tests', '', '']
+        result = subprocess.run(command, capture_output=True, text=True)
+        self.assertEqual(result.returncode,1)
+        self.assertTrue(Path.exists(Path('tests/math.pdf')))
+        self.assertTrue(Path.exists(Path('tests/valid.pdf')))
+
     def test_run_with_path_file_arg(self):
-        # Compile a file using path file argument.  Should coompile both test files.
+        # Compile a file using path file argument.  Should compile both test files.
         remove_glob('tests/*.pdf')          # clean up any existing output artifacts first
         command = [self.python_path, 'entrypoint.py', '', '', 'tests/files.txt']
         result = subprocess.run(command, capture_output=True, text=True)
         self.assertEqual(result.returncode,0)
         self.assertTrue(Path.exists(Path('tests/math.pdf')))
         self.assertTrue(Path.exists(Path('tests/valid.pdf')))
+
+    def test_run_with_multiline_path_file_arg(self):
+        # Compile a file using multiline path file argument.
+        # Should result in a return code of 1 from entrypoint.py.
+        remove_glob('tests/*.pdf')          # clean up any existing output artifacts first
+        command = [self.python_path, 'entrypoint.py', '', '', 'tests/files.txt\ntests/files.txt']
+        result = subprocess.run(command, capture_output=True, text=True)
+        self.assertEqual(result.returncode,1)
+        self.assertFalse(Path.exists(Path('tests/math.pdf')))
+        self.assertFalse(Path.exists(Path('tests/valid.pdf')))
 
     def test_run_with_both_args(self):
         # Compile a file using source file argument and a path file argument.
@@ -66,6 +94,13 @@ class TestAddFunction(unittest.TestCase):
         self.assertEqual(result.returncode,1)
         self.assertFalse(Path.exists(Path('tests/math.pdf')))
         self.assertFalse(Path.exists(Path('tests/valid.pdf')))
+
+    def test_run_with_invalid_file(self):
+        # Compile an invalid source file.
+        # Should result in a return code of 1 from entrypoint.py.
+        command = [self.python_path, 'entrypoint.py', 'tests/bad/invalid.typ', '', '']
+        result = subprocess.run(command, capture_output=True, text=True)
+        self.assertEqual(result.returncode,1)
 
     def test_run_with_options(self):
         # Compile a file using source file argument and an options argument.
